[refactor] Рефакторинг класса MovieQuizViewController
